% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_risk_line_and_label}
\alias{add_risk_line_and_label}
\title{Add Risk Line and Label}
\usage{
add_risk_line_and_label(
  risk_data,
  index,
  total_risks,
  plot_width,
  plot_height,
  size = 3,
  omit_bar = TRUE,
  vertical_spacing = 0.3,
  ...
)
}
\arguments{
\item{risk_data}{Data for a single risk}

\item{index}{Index of the current risk}

\item{total_risks}{Total number of risks in the plot}

\item{plot_width}{Width of the plot}

\item{plot_height}{Height of the plot}

\item{size}{Size of the label text (default is 3)}

\item{omit_bar}{Logical, whether to omit the vertical bar (default is TRUE)}

\item{vertical_spacing}{Vertical spacing between risks (default is 0.3)}

\item{...}{Additional arguments passed to calculate_nudge()}
}
\value{
A list of ggplot2 objects for the risk line and label
}
\description{
This function adds a line and label for a single risk to the magnitude plot.
}
\keyword{internal}
