% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_nudge}
\alias{calculate_nudge}
\title{Calculate Label Nudge}
\usage{
calculate_nudge(
  x_pos,
  y_pos,
  name_length,
  position,
  total_risks,
  plot_width,
  plot_height,
  name_value = 0,
  x_nudge = 0.75,
  y_nudge = 0.4
)
}
\arguments{
\item{x_pos}{X position of the data point}

\item{y_pos}{Y position of the data point}

\item{name_length}{Length of the label name}

\item{position}{Position of the label ("left", "right", or "center")}

\item{total_risks}{Total number of risks in the plot}

\item{plot_width}{Width of the plot}

\item{plot_height}{Height of the plot}

\item{name_value}{Adjustment factor for name length (default is 0)}

\item{x_nudge}{Base nudge in x direction (default is 0.75)}

\item{y_nudge}{Base nudge in y direction (default is 0.4)}
}
\value{
A list with x and y nudge values
}
\description{
This function calculates the nudge values for positioning labels in the magnitude plot.
}
\keyword{internal}
