% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_magnitude_scale}
\alias{create_magnitude_scale}
\title{Create a Magnitude Scale}
\usage{
create_magnitude_scale(min_value, max_value)
}
\arguments{
\item{min_value}{Minimum value for the scale (default is 10)}

\item{max_value}{Maximum value for the scale (default is 10^12)}
}
\value{
A list of ggplot2 objects representing the magnitude scale
}
\description{
This function creates a magnitude scale for use in ggplot2 plots. It generates
a color-coded scale from darkgreen to yellow to darkred, representing values
from 10^0 to 10^12.
}
\keyword{internal}
