% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude_plot.R
\name{magnitude_plot}
\alias{magnitude_plot}
\title{Create a Magnitude Plot}
\usage{
magnitude_plot(
  data,
  value_col,
  name_col,
  font_col = NULL,
  is_log10 = TRUE,
  min_value = 10,
  max_value = 1e+13
)
}
\arguments{
\item{data}{A data frame containing the risk data.}

\item{value_col}{The name of the column containing the values (either in log10 scale or not).}

\item{name_col}{The name of the column containing the risk names.}

\item{font_col}{The name of the column containing the font style (optional).}

\item{is_log10}{Logical, indicating whether the values are already in log10 scale.}

\item{min_value}{The minimum value for the x-axis (default is 1e8).}

\item{max_value}{The maximum value for the x-axis (default is 1e12).}
}
\value{
A ggplot object representing the magnitude plot.
}
\description{
This function creates a magnitude plot in log10 scale to depict cumulative disability adjusted life years (DALYs) lost due to different risks.
}
