% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Magnitude Plot}
\usage{
prepare_data(data, value_col, name_col, font_col, is_log10)
}
\arguments{
\item{data}{A data frame containing the risk data.}

\item{value_col}{The name of the column containing the values.}

\item{name_col}{The name of the column containing the risk names.}

\item{font_col}{The name of the column containing the font style (optional).}

\item{is_log10}{Logical, indicating whether the values are already in log10 scale.}
}
\value{
A data frame with additional columns needed for plotting.
}
\description{
This function prepares the input data for use in the magnitude plot.
}
\keyword{internal}
